requirements.txt
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
Pillow>=9.0.0
gradio>=4.0.0
numpy>=1.24.0
